# harmony-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harmony API",
    "description": "REST API for Harmony.\n\n**Rate Limiting:** All endpoints are subject to rate limiting. When the limit is exceeded, the API responds with `429 Too Many Requests` and a `Retry-After` header indicating when the client may retry.",
    "license": {
      "name": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/v1/servers": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "List all servers the authenticated user is a member of.",
        "description": "# Errors\nReturns `ApiError` on repository error.",
        "operationId": "list_servers",
        "responses": {
          "200": {
            "description": "Server list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Create a new server.",
        "description": "The authenticated user becomes the owner. A default `#general` channel\nand server membership are created automatically.\n\n# Errors\nReturns `ApiError` on validation failure or repository error.",
        "operationId": "create_server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Server created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/servers/{id}": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "Get a server by ID.",
        "description": "# Errors\nReturns `ApiError` if the server is not found or a repository error occurs.",
        "operationId": "get_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Servers"
        ],
        "summary": "Delete a server. Requires owner role.",
        "description": "# Errors\nReturns `ApiError` on insufficient role, not found, or repository error.",
        "operationId": "delete_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Server deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Servers"
        ],
        "summary": "Update a server's name. Requires admin+ role.",
        "description": "# Errors\nReturns `ApiError` on validation failure or repository error.",
        "operationId": "update_server",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Server ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ServerId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateServerRequest": {
        "type": "object",
        "description": "Request body for creating a new server.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Server name (required, non-empty)."
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 Problem Details response.\n\nAll API errors are returned in this standardized format.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "upgrade_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "WHY: Drives upsell when a plan limit is exceeded. The frontend renders\nan \"Upgrade\" CTA using this URL. Only present on `LimitExceeded` errors."
          }
        }
      },
...
      "ServerId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a server (guild).",
        "example": "660e8400-e29b-41d4-a716-446655440001"
      },
      "ServerListResponse": {
        "type": "object",
        "description": "Envelope for a list of servers (ADR-036).",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerResponse"
            }
          }
        }
      },
      "ServerResponse": {
        "type": "object",
        "description": "Server response returned to API consumers.",
        "required": [
          "id",
          "name",
          "ownerId",
          "isDm",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "iconUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/ServerId"
          },
          "isDm": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "$ref": "#/components/schemas/UserId"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateServerRequest": {
        "type": "object",
        "description": "Request body for updating a server.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "New server name (if provided)."
          }
        },
        "additionalProperties": false
      },
...
      "UserId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique identifier for a user (Supabase auth.users UUID).",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT (mobile/API clients)"
      }
    }
  },
...
}